cmdline shenanigans
