+++ mirrord/cli/src/pitm.rs
@@ -38,7 +38,10 @@
 use std::collections::HashMap;
 
 use base64::{Engine, engine::general_purpose::STANDARD};
-use mirrord_layer_lib::process::windows::execution::{LayerManagedProcess, MIRRORD_LAYER_FILE_ENV};
+use mirrord_layer_lib::process::windows::{
+    command_line::build_command_line,
+    execution::{LayerManagedProcess, MIRRORD_LAYER_FILE_ENV},
+};
 use mirrord_progress::NullProgress;
 use serde::Deserialize;
 
@@ -177,78 +180,4 @@ fn build_child_environment(child_env: ChildEnv) -> HashMap<String, String> {
         env.remove(&k);
     }
     env
-}
-
-/// Build a Windows command line string from an executable path plus
-/// arguments, applying CRT-compatible quoting.
-///
-/// `CreateProcessW` does no parsing when `lpApplicationName` is set,
-/// but the child process's own CRT (and anything using
-/// `CommandLineToArgvW`) will still tokenise `lpCommandLine` to
-/// populate `argv`. Convention is therefore that `argv[0]` is the
-/// module name repeated as the first token; some applications rely on
-/// this and some do not, but it costs nothing to be correct.
-fn build_command_line(exe: &str, args: &[String]) -> String {
-    let mut out = String::with_capacity(exe.len() + 2);
-    push_quoted(&mut out, exe);
-    for arg in args {
-        out.push(' ');
-        push_quoted(&mut out, arg);
-    }
-    out
-}
-
-/// Append `arg` to `out`, quoted according to the rules documented for
-/// `CommandLineToArgvW` / the MSVC CRT parser.
-///
-/// Runs of backslashes followed by a literal double quote have to be
-/// doubled, and the quote itself escaped, so that the parser on the
-/// other side reconstructs the original string exactly. A bare arg
-/// with no whitespace or quotes is passed through verbatim; this keeps
-/// the resulting command line readable in the common case.
-fn push_quoted(out: &mut String, arg: &str) {
-    let needs_quoting = arg.is_empty() || arg.chars().any(|c| matches!(c, ' ' | '\t' | '"' | '\n'));
-    if !needs_quoting {
-        out.push_str(arg);
-        return;
-    }
-
-    out.push('"');
-    let mut iter = arg.chars().peekable();
-    while let Some(c) = iter.next() {
-        match c {
-            '\\' => {
-                let mut backslashes: usize = 1;
-                while iter.peek() == Some(&'\\') {
-                    iter.next();
-                    backslashes += 1;
-                }
-                match iter.peek() {
-                    None => {
-                        for _ in 0..backslashes * 2 {
-                            out.push('\\');
-                        }
-                    }
-                    Some(&'"') => {
-                        for _ in 0..backslashes * 2 + 1 {
-                            out.push('\\');
-                        }
-                        out.push('"');
-                        iter.next();
-                    }
-                    Some(_) => {
-                        for _ in 0..backslashes {
-                            out.push('\\');
-                        }
-                    }
-                }
-            }
-            '"' => {
-                out.push('\\');
-                out.push('"');
-            }
-            other => out.push(other),
-        }
-    }
-    out.push('"');
 }
