+++ mirrord/layer-lib/src/process/windows/mod.rs
@@ -3,6 +3,7 @@
 //! This module provides Windows-specific process functionality including
 //! the new modular execution system and synchronization utilities.
 
+pub mod command_line;
 pub mod console;
 pub mod execution;
 pub mod sync;
